Fixing circleci config schema and repeating checkout steps.
It seems that circleci jobs don't share a workspace by default.
There is documentation around shared workspaces, however there's also
a lot of discussion around this, where the general consensus seems to
be confusion about what works and what doesn't.

Better to err on the side of duplicating steps for the moment.