[Resolves #72] Cannot install story branch on windows
# Issue Title

- Cannot install story branch on windows

# Main changes

After merging this PR, the library used to calculate the levenshtein distance is changed. Instead of relying on [levenshtein-ffi](https://github.com/dbalatero/levenshtein-ffi), we now use [damerau-levenshtein](https://github.com/GlobalNamesArchitecture/damerau-levenshtein). This change is due to the fact that levenshtein-ffi fails to install on windows and causes story branch to fail to install.

CHANGELOG
 - Replaced levenshtein distance gem
--- 8< ---